Move accessors to Helpers namespace
