+++ net/DevExtreme.AspNet.Data.Tests/AggregateCalculatorTests.cs
@@ -1,4 +1,5 @@
 ﻿using DevExtreme.AspNet.Data.Aggregation;
+using DevExtreme.AspNet.Data.Helpers;
 using DevExtreme.AspNet.Data.ResponseModel;
 using System;
 using System.Collections.Generic;
