+++ net/DevExtreme.AspNet.Data.Tests/AnonTypeTests.cs
@@ -1,4 +1,5 @@
-﻿using DevExtreme.AspNet.Data.Types;
+﻿using DevExtreme.AspNet.Data.Helpers;
+using DevExtreme.AspNet.Data.Types;
 using System;
 using System.Collections.Generic;
 using System.Linq;
