+++ net/DevExtreme.AspNet.Data.Tests/CustomAggregatorsTests.cs
@@ -1,4 +1,5 @@
 ﻿using DevExtreme.AspNet.Data.Aggregation;
+using DevExtreme.AspNet.Data.Helpers;
 using System;
 using Xunit;
