+++ net/DevExtreme.AspNet.Data.Tests/GroupHelperTests.cs
@@ -1,4 +1,5 @@
-﻿using DevExtreme.AspNet.Data.ResponseModel;
+﻿using DevExtreme.AspNet.Data.Helpers;
+using DevExtreme.AspNet.Data.ResponseModel;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
