+++ net/DevExtreme.AspNet.Data/Aggregation/AvgAggregator.cs
@@ -1,4 +1,5 @@
 ﻿using DevExtreme.AspNet.Data.Aggregation.Accumulators;
+using DevExtreme.AspNet.Data.Helpers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
