+++ net/DevExtreme.AspNet.Data/Helpers/IAccessor.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 
-namespace DevExtreme.AspNet.Data {
+namespace DevExtreme.AspNet.Data.Helpers {
 
     public interface IAccessor<T> {
         object Read(T container, string selector);
