+++ net/DevExtreme.AspNet.Data/RemoteGrouping/RemoteGroupTransformer.cs
@@ -1,4 +1,5 @@
 ﻿using DevExtreme.AspNet.Data.Aggregation;
+using DevExtreme.AspNet.Data.Helpers;
 using DevExtreme.AspNet.Data.ResponseModel;
 using DevExtreme.AspNet.Data.Types;
 using System;
